move fakeServer to dedicated file
